+++ example/main.dart
@@ -1,9 +1,15 @@
 // ignore_for_file: avoid_print
 
+import 'dart:io';
+
 import 'package:dotprompt_dart/dotprompt_dart.dart';
 
 void main() async {
-  final prompt = await DotPrompt.file('example/prompts/greet.prompt');
+  const filename = 'example/prompts/greet.prompt';
+  final prompt = await DotPrompt.stream(
+    File(filename).openRead(),
+    name: filename,
+  );
   final front = prompt.frontMatter; // the model info, settings, etc.
   final result = prompt.render({'name': 'Chris'});